feat: add test count cards
