+++ package.json
@@ -30,7 +30,9 @@
     "sort-by": "^0.0.2"
   },
   "devDependencies": {
+    "@testing-library/dom": "^9.3.3",
     "@testing-library/react": "^14.1.0",
+    "@testing-library/user-event": "^14.5.1",
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
     "@typescript-eslint/eslint-plugin": "^6.8.0",
